Moved out Token class
